+++ src/main/sql/main.sql
@@ -6,5 +6,5 @@ USE vor_tree;
 CREATE TABLE point (
   point_id INTEGER,
   coordinate_index INTEGER,
-  value DOUBLE
+  value LONG
 );
